refactor(evm): split EvmPooledMemory storage into std/zkevm partials
Move the divergent shared-buffer and ZK arena rent/return backends out of
inline #if into EvmPooledMemory.std.cs / .zkevm.cs, and trim comments. This
also drops the mainline-unused System.Numerics using that code-lint flagged
(IDE0005), since BitOperations is now only referenced from the ZK partial.
